Infrastructure as Code (IaC) used for building the infrastructure of the cloud
Simply run the bash script in the folder which uses runs some necessary commands and then uses the YAML file to build the cloud infrastructure in AWS. The YAML files is defined to build the lambda function and add a scheduler to it so that it is invoked once a week at 5pm every Thursday. AWS Cloudformation is then used to create the stack.